requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.5  # For handling file uploads
starlette>=0.25.0  # For CORS middleware
openai>=1.0.0  # For GPT-4 API calls
